module3-task1: * app types place
