--- src/common/interfaces/game/artist-question-answer.interface.ts
@@ -1,6 +1,0 @@
-interface IArtistQuestionAnswer {
-  picture: string;
-  artist: string;
-}
-
-export { IArtistQuestionAnswer };
