--- src/common/interfaces/game/artist-question-song.interface.ts
@@ -1,6 +1,0 @@
-interface IArtistQuestionSong {
-  artist: string;
-  src: string;
-}
-
-export { IArtistQuestionSong };
