+++ src/common/interfaces/game/artist-question.interfaces.ts
@@ -1,0 +1,19 @@
+import { QuestionType } from '~/common/enums/enums';
+
+interface IArtistQuestion {
+  type: QuestionType.ARTIST;
+  song: IArtistQuestionSong;
+  answers: IArtistQuestionAnswer[];
+}
+
+interface IArtistQuestionSong {
+  artist: string;
+  src: string;
+}
+
+interface IArtistQuestionAnswer {
+  picture: string;
+  artist: string;
+}
+
+export { IArtistQuestion, IArtistQuestionSong, IArtistQuestionAnswer };
