+++ src/common/interfaces/game/game.ts
@@ -1,5 +1,2 @@
-export * from './genre-question.interface';
-export * from './genre-question-answer.interface';
-export * from './artist-question.interface';
-export * from './artist-question-song.interface';
-export * from './artist-question-answer.interface';
+export * from './artist-question.interfaces';
+export * from './genre-question.interfaces';
