--- src/common/interfaces/game/genre-question-answer.interface.ts
@@ -1,6 +1,0 @@
-interface IGenreQuestionAnswer {
-  genre: string;
-  src: string;
-}
-
-export { IGenreQuestionAnswer };
