--- src/common/types/callback/binding-cb.type.ts
@@ -1,3 +1,0 @@
-type BindingCb = () => void;
-
-export { BindingCb };
