--- src/common/types/event/change-event.type.ts
@@ -1,5 +1,0 @@
-import * as React from 'react';
-
-type ChangeEvent<T> = React.ChangeEvent<T>;
-
-export { ChangeEvent };
