+++ src/common/types/event/event.ts
@@ -1,3 +1,9 @@
-export * from './change-event.type';
-export * from './form-event.type';
-export * from './input-change-event.type';
+import * as React from 'react';
+
+type ChangeEvent<T> = React.ChangeEvent<T>;
+
+type InputChangeEvent = ChangeEvent<HTMLInputElement>;
+
+type FormEvent = React.FormEvent<HTMLFormElement>;
+
+export { ChangeEvent, InputChangeEvent, FormEvent };
