--- src/common/types/event/form-event.type.ts
@@ -1,5 +1,0 @@
-import * as React from 'react';
-
-type FormEvent = React.FormEvent<HTMLFormElement>;
-
-export { FormEvent };
