--- src/common/types/event/input-change-event.type.ts
@@ -1,5 +1,0 @@
-import { ChangeEvent } from '~/common/types/event/change-event.type';
-
-type InputChangeEvent = ChangeEvent<HTMLInputElement>;
-
-export { InputChangeEvent };
