--- src/common/types/game/answer-artist-question-cb.type.ts
@@ -1,11 +1,0 @@
-import {
-  IArtistQuestion,
-  IArtistQuestionAnswer,
-} from '~/common/interfaces/interfaces';
-
-type AnswerArtistQuestionCb = (
-  question: IArtistQuestion,
-  answer: IArtistQuestionAnswer
-) => void;
-
-export { AnswerArtistQuestionCb };
