--- src/common/types/game/answer-genre-question-cb.type.ts
@@ -1,8 +1,0 @@
-import { IGenreQuestion } from '~/common/interfaces/interfaces';
-
-type AnswerGenreQuestionCb = (
-  question: IGenreQuestion,
-  answers: boolean[]
-) => void;
-
-export { AnswerGenreQuestionCb };
