--- src/common/types/game/change-artist-question-answer-cb.type.ts
@@ -1,5 +1,0 @@
-import { IArtistQuestionAnswer } from '~/common/interfaces/interfaces';
-
-type ChangeArtistQuestionAnswerCb = (answer: IArtistQuestionAnswer) => void;
-
-export { ChangeArtistQuestionAnswerCb };
