--- src/common/types/game/change-genre-question-answer-cb.type.ts
@@ -1,6 +1,0 @@
-type ChangeGenreQuestionAnswerCb = (
-  answerIdx: number,
-  isChecked: boolean
-) => void;
-
-export { ChangeGenreQuestionAnswerCb };
