--- src/common/types/game/game-question.type.ts
@@ -1,8 +1,0 @@
-import {
-  IArtistQuestion,
-  IGenreQuestion,
-} from '~/common/interfaces/interfaces';
-
-type GameQuestion = IGenreQuestion | IArtistQuestion;
-
-export { GameQuestion };
